backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
sqlalchemy==2.0.35
python-multipart==0.0.12
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.1
anthropic==0.40.0
pillow==10.4.0
pypdf==4.3.1
pdf2image==1.17.0
yfinance==0.2.43
httpx==0.27.2
apscheduler==3.10.4
aiofiles==24.1.0
python-telegram-bot==21.6
